requirements.txt
pandas
transformers
sklearn
torch
evaluate
pytorch-crf
dotenv
openai